requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
asyncpraw==7.8.1
pushshift.py>=0.1.2
python-dotenv>=1.0.0
smolagents
pyttsx3>=2.90
soundfile>=0.12.1
numpy>=1.24.3
litellm>=1.10.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
huggingface-hub>=0.28.0
